modularize commands
