+++ .vscode/typescript.code-snippets
@@ -29,4 +29,19 @@
 			"})",
 		],
 	},
+	"Command handler": {
+		"scope": "typescript",
+		"prefix": "cmd",
+		"body": [
+			"import { defineCommand } from \"./helper\";",
+			"",
+			"export default defineCommand({",
+			"\tname: \"${TM_FILENAME_BASE}\",",
+			"\tdescription: \"${2}\",",
+			"\tasync exec(args, channel, user) {",
+			"\t\t$0",
+			"\t},",
+			"})",
+		],
+	},
 }
