+++ src/lib/commands/announce.ts
@@ -1,0 +1,7 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "announce",
+	description: "Call attention to your message with a colored highlight",
+	async exec(args, channel, user) {},
+});
