+++ src/lib/commands/block.ts
@@ -1,0 +1,13 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "block",
+	description: "Block a user from interacting with you on Twitch",
+	args: [
+		{
+			name: "username",
+			required: true,
+		},
+	],
+	async exec(args, channel, user) {},
+});
