+++ src/lib/commands/clear.ts
@@ -1,0 +1,7 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "clear",
+	description: "Clear chat history for non-moderator viewers",
+	async exec(args, channel, user) {},
+});
