+++ src/lib/commands/emote-only.ts
@@ -1,0 +1,7 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "emote-only",
+	description: "Restrict chat to emote only messages",
+	async exec(args, channel, user) {},
+});
