+++ src/lib/commands/follower-only.ts
@@ -1,0 +1,13 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "follower-only",
+	description: "Restrict chat to followers based on their follow duration",
+	args: [
+		{
+			name: "duration",
+			required: false,
+		},
+	],
+	async exec(args, channel, user) {},
+});
