+++ src/lib/commands/marker.ts
@@ -1,0 +1,13 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "marker",
+	description: "Add a stream marker at the current timestamp",
+	args: [
+		{
+			name: "description",
+			required: false,
+		},
+	],
+	async exec(args, channel, user) {},
+});
