+++ src/lib/commands/mods.ts
@@ -1,0 +1,7 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "mods",
+	description: "Display a list of moderators for this channel",
+	async exec(args, channel, user) {},
+});
