+++ src/lib/commands/restrict.ts
@@ -1,0 +1,13 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "restrict",
+	description: "Start restricting a user's messages",
+	args: [
+		{
+			name: "username",
+			required: true,
+		},
+	],
+	async exec(args, channel, user) {},
+});
