+++ src/lib/commands/shared-chat.ts
@@ -1,0 +1,7 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "shared-chat",
+	description: "Start a new shared chat session or join one in an existing collaboration",
+	async exec(args, channel, user) {},
+});
