+++ src/lib/commands/shield.ts
@@ -1,0 +1,7 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "shield",
+	description: "Restrict chat and ban harassing chatters",
+	async exec(args, channel, user) {},
+});
