+++ src/lib/commands/slow.ts
@@ -1,0 +1,13 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "slow",
+	description: "Limit how frequently users can send messages",
+	args: [
+		{
+			name: "duration",
+			required: false,
+		},
+	],
+	async exec(args, channel, user) {},
+});
