+++ src/lib/commands/subscriber-only.ts
@@ -1,0 +1,7 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "subscriber-only",
+	description: "Restrict chat to subscribers only",
+	async exec(args, channel, user) {},
+});
