+++ src/lib/commands/unique.ts
@@ -1,0 +1,7 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "unique",
+	description: "Prevent users from sending duplicate messages",
+	async exec(args, channel, user) {},
+});
