+++ src/lib/commands/unraid.ts
@@ -1,0 +1,7 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "unraid",
+	description: "Stop an ongoing raid",
+	async exec(args, channel, user) {},
+});
