+++ src/lib/commands/vip.ts
@@ -1,0 +1,13 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "vip",
+	description: "Grant VIP status to a user",
+	args: [
+		{
+			name: "username",
+			required: true,
+		},
+	],
+	async exec(args, channel, user) {},
+});
