+++ src/lib/commands/vips.ts
@@ -1,0 +1,7 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "vips",
+	description: "Display a list of VIPs for this channel",
+	async exec(args, channel, user) {},
+});
