+++ src/lib/commands/warn.ts
@@ -1,0 +1,12 @@
+import { defineCommand } from "./helper";
+
+export default defineCommand({
+	name: "warn",
+	description:
+		"Issue a warning to a user that they must acknowledge before sending more messages",
+	args: [
+		{ name: "username", required: true },
+		{ name: "reason", required: true },
+	],
+	async exec(args, channel, user) {},
+});
