+++ src/lib/components/Suggestions.svelte
@@ -6,7 +6,9 @@
 		display: string;
 	}
 
-	export interface CommandSuggestion extends BaseSuggestion, Omit<Required<Command>, "name"> {
+	export interface CommandSuggestion
+		extends BaseSuggestion,
+			Omit<Required<Command>, "name" | "exec"> {
 		type: "command";
 	}
